+++ src/services/Matrix.php
@@ -526,8 +526,8 @@ class Matrix extends Component
                 }
             }
 
-            // Only make block type changes if we're not in the middle of applying YAML changes
-            if (!Craft::$app->getProjectConfig()->getIsApplyingYamlChanges()) {
+            // Only make block type changes if we're not in the middle of applying external changes
+            if (!Craft::$app->getProjectConfig()->getIsApplyingExternalChanges()) {
                 // Delete the old block types first, in case there's a handle conflict with one of the new ones
                 $oldBlockTypes = $this->getBlockTypesByFieldId($matrixField->id);
                 $oldBlockTypesById = [];